+++ src/types/policy.types.ts
@@ -133,7 +133,7 @@ export interface NewPolicyForm {
   carrierId: string;
   productId?: string;
   product: ProductType;
-  submitDate?: string;
+  submitDate: string;
   effectiveDate: string;
   expirationDate?: string;
   termLength?: number;
@@ -195,7 +195,7 @@ export interface CreatePolicyData {
   productId?: string;
   userId: string;
   product: ProductType;
-  submitDate?: Date;
+  submitDate: Date;
   effectiveDate: Date;
   termLength?: number;
   expirationDate?: Date;
@@ -255,8 +255,12 @@ export function createPolicyDataToInsert(data: CreatePolicyData): PolicyInsert {
     policy_number: data.policyNumber,
     client_id: data.clientId,
     carrier_id: data.carrierId,
+    product_id: data.productId,
     user_id: data.userId,
     product: data.product,
+    submit_date: data.submitDate
+      ? data.submitDate.toISOString().split("T")[0]
+      : data.effectiveDate.toISOString().split("T")[0],
     effective_date: data.effectiveDate.toISOString().split("T")[0],
     term_length: data.termLength,
     expiration_date: data.expirationDate?.toISOString().split("T")[0],